wip deps use my forked flutter_local_notifications, with collision fix
(The alternative workaround would be to pass distinct notif "id" integers.
These could be a hash of the tag... but if it's merely hashCode,
there'd be risk of collisions.)

I feel like the right answer here is that the plugin should be a much
thinner layer, much closer to just exposing the underlying API.
(Also it can use Pigeon to simplify the transport boilerplate.)
Then our Dart code can express what we want.