WIP try connecting stdout/stderr while running
I'm not finding a way in the Dart SDK to do this the *actual* desired
way, where the child process shares the file description via fork.

Here's an attempt at second best.  See:
  https://github.com/dart-lang/sdk/issues/ 44573
  https://github.com/dart-lang/sdk/issues/ 5607
  https://stackoverflow.com/questions/72183086/dart-dont-buffer-process-stdout-tell-process-that-it-is-running-from-a-termina

But this fails in a new and exciting way -- the script seems to just
silently exit while awaiting its child!