wip aha, do that second-best wiring correctly
The `Stream.pipe` version *closes* the downstream when the
upstream finishes.  So once the first check completed, we
wouldn't see any further output, including whatever error
was caused by that.

This was cleared up by simply reading the `Stream.pipe` docs.

Still distinctly second-best, particularly in that the child processes
don't realize they're talking to a terminal (when they are) and so
spew voluminous output better suited to a CI log.