wipref db: Fix drift_schema_v2.json to match current schema
This change happens when rerunning the `dart run drift_dev`
commands prescribed in the comments at [AppDatabase.schemaVersion].
The discrepancy appears to have been there since this schema file
was introduced in commit 066fe259a, oops.

After pretty-printing (with `jq`), the diff looks like this:

    @@ -38,3 +38,7 @@ "name": "realm_url",
                 "default_client_dart": null,
    -            "dsl_features": []
    +            "dsl_features": [],
    +            "type_converter": {
    +              "dart_expr": "const UriConverter()",
    +              "dart_type_name": "Uri"
    +            }
               },

So it reflects the use of UriConverter here:

    Column<String> get realmUrl => text().map(const UriConverter())();

Probably what happened is that I generated the schema file while
on a draft of that branch that didn't yet use UriConverter there;
then added it and didn't realize it would affect the schema file.

We'll shortly have an omnibus run-all-checks script which in
particular reruns those files and checks that nothing changed;
and we'll then run that script in CI (60), which will prevent any
discrepancies like this one from creeping in in the future.
A draft of that script is how I discovered this.