wip dedupe reload on poll error; and on non-transient request errors, reload rather than retry
Hmm or maybe for those non-transient request errors we should stick
with retrying?  Maybe up to a bounded number of times?

Those are by definition a bug either in the server or in our
formulating the request: it's either a 4xx we didn't expect,
or a malformed response (so in either case, a mismatch of our
expectations to the server's behavior, and therefore a bug in
one or the other), or something other than ApiRequestException
entirely (and so a bug in the API request code).

It's therefore hard to predict what's most likely to work.

Maybe the most likely cause is that the server is having operational
trouble that's not showing up as 5xx responses.

... Hmm, do we handle rate-limit errors?  No, we don't.
So *that's* one possible cause, anyway.  Certainly we should
back off if we get one of those!  ... Though I'm not sure
that getEvents is actually currently subject to rate-limiting,
being in Tornado rather than Django.

Also, when we abandon retrying getEvents and switch to re-register,
we do end up backing off if that fails.  It's just that that's a
heavier request in the first place.