test: Use stream ID in default name and description in eg.stream
In particular this means that repeated calls `eg.stream()` with
no arguments will get distinct names, as well as IDs, so that they
can all be added to the same store without collision.

Then also simplify away the various places where a test was passing
a stream name when the only thing about the name that mattered to
the test was being distinct from other names (or when nothing about
the name mattered to the test at all).