autocomplete: Carry whole User object on UserMentionAutocompleteResult
This is very nearly NFC; we use the User object where we previously
would have used the user ID to look up a User object in the store.
As long as it's the same PerAccountStore object involved, that would
have given us the identical User object we return here; so even if the
user's name or avatar etc. has been updated by some concurrent event,
both ways will get us the same updated values.

The one way this can affect behavior is if the ComposeAutocomplete
widget which consumes these result objects has a different ambient
PerAccountStore than the one the results were computed with.  I
believe that can happen if the store has just gotten replaced, e.g.
because the event queue on the server expired, and will last until
the search completes on the new store.  Even then, the new behavior
seems more correct: e.g. if the user's name has changed so they no
longer match the query, then it'd look glitchy to be showing their
new name even though it doesn't match.

On the other hand this change is useful because it will help us
simplify the AutocompleteView interface by using the result objects to
also describe the candidates.