notif test: Cut debugEnableRegisterNotificationToken in favor of calling dispose
For these tests that call UpdateMachine.load and therefore
incidentally cause UpdateMachine.registerNotificationToken to get
called, the latter is actually harmless in itself: these tests don't
set a notification token, so the register call has no effect.

The problem that merely removing the flag would cause shows up in
later tests in this file, which do set a token.  The reason that's a
problem is that these tests were leaking state -- their update
machines were remaining active, listening for notification token
updates and trying to make requests to register those.

The right fix is to make the tests stop leaking that state, by
calling `dispose` on the update machines.  Then we no longer need
this debug flag.