logo: Fix nondeterminism in ImageMagick step
The ImageMagick `convert` command, introduced in b13cf4e78,
apparently inserts the current time into the file.  That makes the
build not reproducible.

Fix that using the "strip-nondeterminism" tool from the
Reproducible Builds project:
  https://reproducible-builds.org/docs/timestamps/

The resulting change to the affected file can be seen using `sng`,
and purely removes a pair of timestamps:

$ f=ios/Runner/Assets.xcassets/AppIcon.appiconset/Icon-1024x1024@1x.png \
  && diff -U3 <(git show @:"$f" | sng) \
              <(<"$f" sng)
--- /dev/fd/63	2025-01-06 21:11:42.422900540 -0800
+++ /dev/fd/62	2025-01-06 21:11:42.426900539 -0800
@@ -13,14 +13,6 @@
 bKGD {red: 255;  green: 255;  blue: 255;}
 pHYs {xpixels: 161260; ypixels: 161260; per: meter;}  # (4096 dpi)
 tEXt {
-    keyword: "date:create";
-    text: "2023-11-27T17:11:13-08:00";
-}
-tEXt {
-    keyword: "date:modify";
-    text: "2023-11-27T17:11:13-08:00";
-}
-tEXt {
     keyword: "Software";
     text: "www.inkscape.org";
 }