wip set haveNewest from response; TODO fact-check that server is reliable on this
Because we set `anchor=newest`, the server should be setting this to
true even though we set `num_after=0`.

And empirically it does do so, in a quick test on CZO.

This is useful, though, because it will let us test the logic
downstream of haveNewest, by setting up bogus responses to the fetch
request which ignore that it was actually for `anchor=newest`,
even while that logic is effectively unreachable in the live app
(with a well-behaved server).